[web] Move product selection to products components
