+++ web/src/components/product/ProductSelectionPage.jsx
