+++ web/src/index.js
@@ -37,7 +37,8 @@ import App from "~/App";
 import Main from "~/Main";
 import DevServerWrapper from "~/DevServerWrapper";
 import { Overview } from "~/components/overview";
-import { ProductSelectionPage, SoftwarePage } from "~/components/software";
+import { ProductSelectionPage } from "~/components/product";
+import { SoftwarePage } from "~/components/software";
 import { ProposalPage as StoragePage, ISCSIPage, DASDPage, ZFCPPage } from "~/components/storage";
 import { UsersPage } from "~/components/users";
 import { L10nPage } from "~/components/l10n";
